fix: crossfade underline visibility, integer T1/T2 vmin/vmax display
- _crossfading guard prevents underline from reappearing during mode-transition callback
- T1/T2 vmin/vmax use int() so values are true integers (no .00 suffix)
- formatValueGroup detects integer-like values and uses 0 minDecimals when appropriate

+++ src/arrayview/_synthetic_mri.py
@@ -150,7 +150,7 @@ def _qmri_adjust_vmin_vmax(vmin: float, vmax: float, role: str) -> tuple[float, 
     if not role:
         return vmin, vmax
     if role in ("t1", "t2"):
-        return float(np.floor(vmin)), float(np.ceil(vmax))
+        return float(int(np.floor(vmin))), float(int(np.ceil(vmax)))
     if role == "pd":
         return 0.0, vmax
     if role == "b1":
